+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog of lizard-map
 2.1.3 (unreleased)
 ------------------
 
+- Made the adapter.extent function optional. In
+  WorkspaceItem.has_extent we now just detect if the function is at
+  all present.  #3041, #3036.
+
 - Added wms layers to transparency function. #3039.
 
 - Updated googlemaps_api_keys in fixture: the new key is for